# optimize uses of count_while