The `count_while f >>= advance` pattern was a signifcant source of parser allocations. Elminiate that pattern by passing a `with_buffer` function to count_while, which will be passed a view into the input that count_while will consume. The result of `with_buffer` is now the result of the parser.

This pull request also adds some looping benchmarks that were used to evaluate the change, together with existing json benchmarks.

## Loops
Before:
```
┌──────────────────┬──────────┬────────────┬──────────┬──────────┬───────────┬──────────┬────────────┐
│ Name             │ Time/Run │    mWd/Run │ mjWd/Run │ Prom/Run │   mGC/Run │ mjGC/Run │ Percentage │
├──────────────────┼──────────┼────────────┼──────────┼──────────┼───────────┼──────────┼────────────┤
│ skip_while  true │   1.92us │     23.00w │          │          │   0.09e-3 │          │      5.00% │
│ take_while  true │   2.00us │    154.00w │          │          │   0.59e-3 │          │      5.19% │
│ take_while1 true │   1.98us │    154.00w │          │          │   0.59e-3 │          │      5.15% │
│ many any_char    │  38.45us │ 27_708.62w │  861.59w │  861.59w │ 107.95e-3 │  4.51e-3 │    100.00% │
└──────────────────┴──────────┴────────────┴──────────┴──────────┴───────────┴──────────┴────────────┘
```
After:
```
┌──────────────────┬──────────┬────────────┬──────────┬──────────┬───────────┬──────────┬────────────┐
│ Name             │ Time/Run │    mWd/Run │ mjWd/Run │ Prom/Run │   mGC/Run │ mjGC/Run │ Percentage │
├──────────────────┼──────────┼────────────┼──────────┼──────────┼───────────┼──────────┼────────────┤
│ skip_while  true │   1.97us │      9.00w │          │          │   0.03e-3 │          │      5.04% │
│ take_while  true │   2.01us │    139.00w │          │          │   0.53e-3 │          │      5.15% │
│ take_while1 true │   2.01us │    139.00w │          │          │   0.53e-3 │          │      5.15% │
│ many any_char    │  39.06us │ 27_708.62w │  859.56w │  859.56w │ 107.96e-3 │  4.51e-3 │    100.00% │
└──────────────────┴──────────┴────────────┴──────────┴──────────┴───────────┴──────────┴────────────┘
```

## JSON

Before:
```
┌─────────────┬─────────────┬────────────┬─────────────┬─────────────┬──────────────┬───────────┬────────────┐
│ Name        │    Time/Run │    mWd/Run │    mjWd/Run │    Prom/Run │      mGC/Run │  mjGC/Run │ Percentage │
├─────────────┼─────────────┼────────────┼─────────────┼─────────────┼──────────────┼───────────┼────────────┤
│ twitter1    │     15.87us │     7.14kw │       9.64w │       9.64w │     27.29e-3 │   0.09e-3 │      0.14% │
│ twitter10   │    108.80us │    44.22kw │     255.19w │     255.19w │    169.65e-3 │   1.96e-3 │      0.94% │
│ twitter20   │    213.00us │    87.34kw │     714.45w │     714.45w │    335.50e-3 │   4.60e-3 │      1.84% │
│ twitter-big │ 11_576.81us │ 3_482.93kw │ 190_347.26w │ 190_347.26w │ 13_776.24e-3 │ 844.32e-3 │    100.00% │
└─────────────┴─────────────┴────────────┴─────────────┴─────────────┴──────────────┴───────────┴────────────┘
```
After:
```
┌─────────────┬─────────────┬────────────┬─────────────┬─────────────┬─────────────┬───────────┬────────────┐
│ Name        │    Time/Run │    mWd/Run │    mjWd/Run │    Prom/Run │     mGC/Run │  mjGC/Run │ Percentage │
├─────────────┼─────────────┼────────────┼─────────────┼─────────────┼─────────────┼───────────┼────────────┤
│ twitter1    │     14.25us │     4.92kw │       6.49w │       6.49w │    18.81e-3 │   0.06e-3 │      0.13% │
│ twitter10   │     99.29us │    30.52kw │     174.91w │     174.91w │   117.12e-3 │   1.35e-3 │      0.88% │
│ twitter20   │    192.65us │    59.56kw │     651.76w │     651.76w │   229.27e-3 │   4.22e-3 │      1.71% │
│ twitter-big │ 11_252.43us │ 2_381.33kw │ 184_363.54w │ 184_363.54w │ 9_467.42e-3 │ 701.74e-3 │    100.00% │
└─────────────┴─────────────┴────────────┴─────────────┴─────────────┴─────────────┴───────────┴────────────┘
```